add data layer and unit test

+++ app/src/main/java/com/harrycampaz/searchproducts/domain/entities/ProductEntity.kt
@@ -5,5 +5,5 @@ data class ProductEntity(
     val title: String,
     val price: Double,
     val thumbnail: String,
-    val available_quantity: Double,
+    val available_quantity: Int,
 )
